backend/scraper/requirements.txt
# Core scraping libraries
scrapy==2.11.0          # Powerful, async scraping framework
playwright==1.40.0      # Handle JavaScript-heavy pages
beautifulsoup4==4.12.2  # HTML parsing for simple cases
lxml==4.9.3            # Fast XML/HTML processing

# Request handling
httpx==0.25.2          # Modern async HTTP client
requests==2.31.0       # Fallback for sync operations
urllib3==2.1.0         # URL parsing and handling
fake-useragent==1.4.0  # Rotate user agents to avoid detection

# Data processing
pandas==2.1.4          # Data manipulation and analysis
numpy==1.26.2          # Numerical operations
pydantic==2.5.2        # Data validation and serialization

# Database
psycopg2-binary==2.9.9 # PostgreSQL adapter
sqlalchemy==2.0.23     # ORM for database operations
alembic==1.13.1        # Database migrations

# GCP integrations
google-cloud-storage==2.13.0    # Store raw scraped data
google-cloud-pubsub==2.19.0     # Message queue for tasks
google-cloud-logging==3.8.0     # Centralized logging
google-cloud-secret-manager==2.17.0  # Secure credential storage

# Caching and queue
redis==5.0.1           # Caching and distributed locks
celery==5.3.4          # Distributed task queue

# Monitoring and debugging
sentry-sdk==1.39.1     # Error tracking
prometheus-client==0.19.0  # Metrics export
structlog==23.2.0      # Structured logging

# Utilities
python-dotenv==1.0.0   # Environment variable management
tenacity==8.2.3        # Retry logic with exponential backoff
schedule==1.2.0        # Simple job scheduling
pytz==2023.3           # Timezone handling